figured out how to split out gui node functionality into multiple files
